First pass of Player Movement and Camera Movement
On the PlayerMovement script, there is a sliding variable (currently 1 or 2) which dictates which movement to use. Currently 1 is set to movement relative to the player only and 2 is set to movement relative to the camera. Camera settings are not perfect, simply used a Cinemachine free move - adjust settings as you feel necessary, can be found in the Prefab folder.  Working with both keyboard/mouse and controller (only PS4 tested).

Update syntax