Add Combo Detection plus bonus Team and Collectable Update
Does not yet include the actual digging. Will be tackled in the following issue! 

* Update Collectables and Teams

To not use event based programming, which was causing issues with our scalability :(

* Focus on Triangle Combo

* Add more work on Collectables and Teams

* Complete Digging Combo Detection

* Make the Triangular Comboed Tiles more Triangular

* Make the Linear Comboed Tiles Rectangular

* Fix Triangular Edge Case

When A B C are in a line but can form a triangle, no tiles could have been selected, making 2 line combos more desirable. 

Instead, now remove the Line combos that overlap these players, add their tiles to the triangle combo, and make the triangle combo always as good or better than the line combo

* Adjust for PR review

* Fix Line Rectangle Detection

Main issue was 45o instead of 90o, 
so the area was incorrectly slanted

* More PR Review